requirements.txt
asgiref==3.9.1
dj-database-url==2.0.0
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==21.2.0
packaging==25.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.6.0
python-dotenv==1.0.0
